[8.0][FIX][web_list_view_sequence]
Steps to reproduce: Choose a Many2one field in a form in edit mode, and then select the "Search more" option. When the search view opens, column headers are not well displayed, and rownumber is in under the label of the first field, and first field data is under second field label, and so on.
This fix solves this issue.